ipmitool: Add cipher suite 17 as default
Sets ipmitool's cipher suite to 17.

After upgrading Lenovo's TSM firmware to latest on our RD350 servers
we saw that the default crypto algorithm used by ipmitool to communicate with
IPMI, is not supported anymore.
By manualy selecting the 17th algorithm communication is enstablished.
This issue is fixed in newer ipmitool versions, but are not yet introduced
for Debian buster/bullseye variants.